+++ cses/introductoryProblems/missingNumber.java
@@ -5,11 +5,11 @@ public class missingNumber
 	public static void main(String[] args)
 	{
 		Scanner sc = new Scanner(System.in);
-		int n = sc.nextInt();
-		int sum = n * (n + 1)/2;
-		for (int i = 0; i < (n - 1); i++)
+		long n = sc.nextLong();
+		long sum = n * (n + 1)/2;
+		for (long i = 0; i < (n - 1); i++)
 		{
-			int num = sc.nextInt();
+			long num = sc.nextLong();
 			sum = sum - num;
 		}
 		System.out.println(sum);
